requirements.txt
pywavefront >=1.3.0,<1.4.0
PyQt5 >=5.13.1,<5.14
PyOpenGL >=3.1.0,<3.2
piglet